Correcao Regra de negocio agendamento
